# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Muse API",
    "description": "Muse Music Streaming Service API Documentation",
    "contact": {
      "email": "mubareksd@gmail.com"
    },
    "license": {
      "name": "GPL 3.0",
      "url": "https://www.gnu.org/licenses/gpl-3.0.html"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://api.muse.com/v1"
    }
  ],
...
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "tags": ["Authentication"],
        "operationId": "refresh",
        "summary": "Refresh",
        "description": "Refresh access token",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Refresh"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refreshed Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/me/playlists": {
      "get": {
        "tags": ["Me"],
        "summary": "Me",
        "description": "Get Current users Playlists",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched Successfully"
          }
        }
      }
    },
...
    "/playlists": {
      "get": {
        "tags": ["Playlists"],
        "summary": "Playlists",
        "description": "Get Several Playlists",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "ids",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "limit",
            "example": 10,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "offset",
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched Successfully"
          }
        }
      }
    },
    "/Playlists/{id}": {
      "get": {
        "tags": ["Playlists"],
        "summary": "Playlists",
        "description": "Get Playlist with id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "example": "5272457a980bab8465e22ca2aef2ef482a199a237923",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched Successfully"
          }
        }
      },
      "put": {
        "tags": ["Playlists"],
        "summary": "Playlists",
        "description": "Update Playlist with id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "example": "5272457a980bab8465e22ca2aef2ef482a199a237923",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Playlists"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated Successfully"
          }
        }
      }
    },
...
    "/users/{id}/playlists": {
      "get": {
        "tags": ["Users"],
        "summary": "Users",
        "description": "Get User's Playlists with id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "example": "5272457a980bab8465e22ca2aef2ef482a199a237923",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Fetched Successfully"
          }
        }
      },
      "post": {
        "tags": ["Users"],
        "summary": "Users",
        "description": "Create User's Playlist with id",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "id",
            "example": "5272457a980bab8465e22ca2aef2ef482a199a237923",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Playlists"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created Successfully"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Refresh": {
        "required": ["refreshToken"],
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJlMjQyYmM5YTBjZWVhODUyNDAwZDNmOTI3ZjA3NTU3ZTRhMzZiOWYzNGU3ZiIsInJvbGUiOiJBUlRJU1QiLCJpYXQiOjE2ODcxNTMxMTAsImV4cCI6MTY4OTc0NTExMH0.SgNOHhdb7ooVkgWlKMEX_IwfetR2vIK4aiDU7EsTXlU",
            "description": "User Refresh Token"
          }
        }
      },
      "RefreshResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VySWQiOiJlMjQyYmM5YTBjZWVhODUyNDAwZDNmOTI3ZjA3NTU3ZTRhMzZiOWYzNGU3ZiIsInJvbGUiOiJBUlRJU1QiLCJpYXQiOjE2ODcxNTMxMTAsImV4cCI6MTY4NzE1NjcxMH0.8RnciaKJQI8CZcYYe2wJoaK9NJPHNnB--QdTgFlp85A",
            "description": "User Access Token"
          }
        }
      },
...
    },
...
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}